feat(extension): fixing some bugs on app locking and some refactory
