--- apps/extension/src/App/LockWrapper/index.ts
@@ -1,1 +1,0 @@
-export * from "./LockWrapper";
