+++ packages/components/src/LinkButton/LinkButton.components.ts
@@ -1,9 +1,18 @@
-import { ThemeColor, color, fontSize, spacement } from "@namada/utils";
+import {
+  FontSize,
+  ThemeColor,
+  color,
+  fontSize,
+  spacement,
+} from "@namada/utils";
 import styled from "styled-components";
 
-export const LinkButtonContainer = styled.div<{ themeColor: ThemeColor }>`
+export const LinkButtonContainer = styled.div<{
+  themeColor: ThemeColor;
+  size?: keyof FontSize;
+}>`
   color: ${(props) => color(props.themeColor, "main")(props)};
-  font-size: ${fontSize("base")};
+  font-size: ${(props) => fontSize(props.size || "base")(props)}};
   padding-bottom: ${spacement(1)};
   text-align: center;
   display: inline-block;
